+++ src/diffusers/pipelines/flux2/pipeline_flux2_klein.py
@@ -279,9 +279,7 @@ class Flux2KleinPipeline(DiffusionPipeline, Flux2LoraLoaderMixin):
             coords = torch.cartesian_prod(t, h, w, l)
             out_ids.append(coords)
 
-        # Cast position ids to float32: these are RoPE coordinate indices and the Neuron
-        # compiler does not support int64 tensors. float32 is exact for this index range.
-        return torch.stack(out_ids).float()
+        return torch.stack(out_ids)
 
     @staticmethod
     # Copied from diffusers.pipelines.flux2.pipeline_flux2.Flux2Pipeline._prepare_latent_ids
@@ -314,9 +312,7 @@ class Flux2KleinPipeline(DiffusionPipeline, Flux2LoraLoaderMixin):
         # Expand to batch: (B, H*W, 4)
         latent_ids = latent_ids.unsqueeze(0).expand(batch_size, -1, -1)
 
-        # Cast position ids to float32: these are RoPE coordinate indices and the Neuron
-        # compiler does not support int64 tensors. float32 is exact for this index range.
-        return latent_ids.float()
+        return latent_ids
 
     @staticmethod
     # Copied from diffusers.pipelines.flux2.pipeline_flux2.Flux2Pipeline._prepare_image_ids
@@ -367,9 +363,7 @@ class Flux2KleinPipeline(DiffusionPipeline, Flux2LoraLoaderMixin):
         image_latent_ids = torch.cat(image_latent_ids, dim=0)
         image_latent_ids = image_latent_ids.unsqueeze(0)
 
-        # Cast position ids to float32: these are RoPE coordinate indices and the Neuron
-        # compiler does not support int64 tensors. float32 is exact for this index range.
-        return image_latent_ids.float()
+        return image_latent_ids
 
     @staticmethod
     # Copied from diffusers.pipelines.flux2.pipeline_flux2.Flux2Pipeline._patchify_latents
